requirements.txt
streamlit>=1.32.0
openai
pydantic>=2.0.0
aiohttp>=3.9.0
asyncio>=3.4.3
python-docx>=0.8.11
PyPDF2>=3.0.0 